+++ ignite/templates/query/placeholders.go
@@ -1,3 +1,0 @@
-package query
-
-const PlaceholderAutoCLIQuery = "// this line is used by ignite scaffolding # autocli/query"