+++ changelog.md
@@ -2,18 +2,27 @@
 
 ## Unreleased
 
+### Changes
+
+- [#4901](https://github.com/ignite/cli/pull/4901) Enable automatic switching to daemon mode when not a terminal.
+
+## [`v29.9.0`](https://github.com/ignite/cli/releases/tag/v29.9.0)
+
 ## Features
 
 - [#4877](https://github.com/ignite/cli/pull/4877) Remove app config and ibc add route placeholders.
+- [#3094](https://github.com/ignite/cli/issues/3094) Add support for arrays with custom types.
 
 ### Changes
 
 - [#4880](https://github.com/ignite/cli/pull/4880) Add possibility to scaffold underscore package.
 - [#4878](https://github.com/ignite/cli/pull/4878) Improve the `xast` package readability.
 - [#4883](https://github.com/ignite/cli/pull/4883) Remove ibc keys placeholders.
+- [#4884](https://github.com/ignite/cli/pull/4884) Remove autocli placeholders.
 
 ### Fixes
 
+- [#4886](https://github.com/ignite/cli/pull/4886) Fix chain scaffolding checks.
 - [#4889](https://github.com/ignite/cli/pull/4889) Plugin data race.
 
 ## [`v29.8.0`](https://github.com/ignite/cli/releases/tag/v29.8.0)